# back/transport-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  leads.addLeadRequest:
    properties:
      consumer_id:
        type: integer
      price:
        type: integer
      producer_id:
        type: integer
      responsible_id:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
  leads.leadData:
    properties:
      consumer_id:
        type: integer
      created_at:
        type: integer
      lead_id:
        type: integer
      price:
        type: integer
      producer_id:
        type: integer
      responsible_id:
        type: integer
      status:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
  leads.leadListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/leads.leadData'
        type: array
      error:
        type: string
    type: object
  leads.leadResponse:
    properties:
      data:
        $ref: '#/definitions/leads.leadData'
      error:
        type: string
    type: object
  leads.stageResponse:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
  leads.updateLeadRequest:
    properties:
      price:
        type: integer
      responsible_id:
        type: integer
      status:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
  tasks.addTaskRequest:
    properties:
      deadline:
        type: integer
      responsible_id:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
  tasks.stageResponse:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
  tasks.taskData:
    properties:
      company_id:
        type: integer
      created_at:
        type: integer
      created_id:
        type: integer
      deadline:
        type: integer
      responsible_id:
        type: integer
      stage:
        type: integer
      task_id:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
  tasks.taskListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/tasks.taskData'
        type: array
      error:
        type: string
    type: object
  tasks.taskResponse:
    properties:
      data:
        $ref: '#/definitions/tasks.taskData'
      error:
        type: string
    type: object
  tasks.updateTaskRequest:
    properties:
      deadline:
        type: integer
      responsible_id:
        type: integer
      stage:
        type: integer
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Swagger документация к API
  title: BRM API
  version: "1.0"
paths:
...
  /leads:
    get:
      description: Получает список сделок с использованием фильтрации и пагинации
      parameters:
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: Поиск по названию/тексту
        in: query
        name: name
        required: true
        type: string
      - description: Поиск по этапу
        in: query
        name: stage
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение сделок
          schema:
            $ref: '#/definitions/leads.leadListResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/leads.leadListResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/leads.leadListResponse'
      summary: Получение списка сделок
      tags:
      - leads
    post:
      consumes:
      - application/json
      description: Добавляет новую сделку
      parameters:
      - description: Новая сделка в JSON
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/leads.addLeadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное добавление задачи
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/leads.leadResponse'
      summary: Добавление новой сделки
      tags:
      - leads
  /leads/{id}:
    delete:
      description: Безвозвратно удаляет сделку
      parameters:
      - description: id сделки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление сделки
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "404":
          description: Сделка не найдена
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/leads.leadResponse'
      summary: Удаление сделки
      tags:
      - leads
    get:
      description: Получает сделку по id
      parameters:
      - description: id сделки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение сделки
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "404":
          description: Задача не найдена
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/leads.leadResponse'
      summary: Получение сделки
      tags:
      - leads
    put:
      consumes:
      - application/json
      description: Изменяет одно или несколько полей сделки
      parameters:
      - description: id сделки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые поля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/leads.updateLeadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление данных
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "404":
          description: Сделка не найдена
          schema:
            $ref: '#/definitions/leads.leadResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/leads.leadResponse'
      summary: Редактирование сделки
      tags:
      - leads
  /leads/stages:
    get:
      description: Возвращает словарь из этапов и их id
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение данных
          schema:
            $ref: '#/definitions/leads.stageResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/leads.stageResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/leads.stageResponse'
      summary: Получение этапов сделки
      tags:
      - leads
  /tasks:
    get:
      description: Получает список задач с использованием фильтрации и пагинации
      parameters:
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: Поиск по названию/тексту
        in: query
        name: name
        required: true
        type: string
      - description: Поиск по этапу
        in: query
        name: stage
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение задач
          schema:
            $ref: '#/definitions/tasks.taskListResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/tasks.taskListResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/tasks.taskListResponse'
      summary: Получение списка задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Добавляет новую задачу
      parameters:
      - description: Новая задача в JSON
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/tasks.addTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное добавление задачи
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/tasks.taskResponse'
      summary: Добавление новой задачи
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Безвозвратно удаляет задачу
      parameters:
      - description: id задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление задачи
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "404":
          description: Задача не найдена
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/tasks.taskResponse'
      summary: Удаление задачи
      tags:
      - tasks
    get:
      description: Получает задачу по id
      parameters:
      - description: id задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение задачи
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "404":
          description: Задача не найдена
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/tasks.taskResponse'
      summary: Получение задачи
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Изменяет одно или несколько полей задачи
      parameters:
      - description: id задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Новые поля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/tasks.updateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление данных
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "404":
          description: Задача не найдена
          schema:
            $ref: '#/definitions/tasks.taskResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/tasks.taskResponse'
      summary: Редактирование задачи
      tags:
      - tasks
  /tasks/stages:
    get:
      description: Возвращает словарь из этапов и их id
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение данных
          schema:
            $ref: '#/definitions/tasks.stageResponse'
        "400":
          description: Неверный формат входных данных
          schema:
            $ref: '#/definitions/tasks.stageResponse'
        "500":
          description: Проблемы на стороне сервера
          schema:
            $ref: '#/definitions/tasks.stageResponse'
      summary: Получение этапов задачи
      tags:
      - tasks
swagger: "2.0"